requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
spacy==3.7.2
pandas==2.1.3
lxml==4.9.3
fake-useragent==1.4.0
schedule==1.2.0
python-dateutil==2.8.2
